requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.10
greenlet==1.1.2
idna==3.3
numpy==1.22.1
pandas==1.3.5
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.29
toml==0.10.2
urllib3==1.26.8
